+++ ChangeLog.md
@@ -16,6 +16,9 @@ All notable changes to this project will be documented in this file.
 - :openai module support for assistant message api (in work)
 - :openai module support for assistant run api (...)
 - :openai module support for assistant run steps api (...)
+- support for shebang lines. Venice allows shebang lines simply by making `#!` 
+  equivalent to `;` comment
+- support for `#_` read macro to skip forms